services/agent/requirements.txt
# ---- Core API ----
fastapi==0.115.6
uvicorn==0.30.6

# ---- Config ----
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1

# ---- MongoDB ----
pymongo==4.10.1

# ---- LangChain core ----
langchain-core>=0.3.65
langchain-community==0.3.25
langgraph==0.2.56

# ---- Google Gemini ----
langchain-google-genai==2.0.8
google-generativeai==0.8.3

# ---- Vector DB (Chroma) ----
chromadb==0.5.20

# ---- Embeddings ----
sentence-transformers==3.3.1
torch>=2.0.0

# ---- File handling ----
python-multipart==0.0.9

# ---- JWT Authentication ----
PyJWT==2.9.0